requirements.txt
langchain
langchain-core
langchain-community
langchain-astradb
datasets
pypdf
python-dotenv
flask
requests
transformers
huggingface_hub
-e .
